+++ tests/query_constraint_parity.rs
@@ -75,7 +75,7 @@ where
     };
 
     assert_engine("sequential", make_query().sequential().collect());
-    assert_engine("ordinary residual-default", make_query().collect());
+    assert_engine("ordinary feature-selected", make_query().collect());
     assert_engine("lazy DAG", make_query().solve_dag_lazy().collect());
 
     let mut lazy = make_query().solve_dag_lazy();
@@ -101,7 +101,7 @@ where
     #[cfg(feature = "parallel")]
     {
         assert_engine(
-            "ordinary parallel residual",
+            "ordinary parallel feature-selected",
             make_query().into_par_iter().collect::<Vec<_>>(),
         );
         assert_engine(